add bulk delete to IUserService

+++ src/Core/Services/Implementations/UserService.cs
@@ -32,6 +32,8 @@ using Microsoft.Extensions.Options;
 using File = System.IO.File;
 using JsonSerializer = System.Text.Json.JsonSerializer;
 
+#nullable enable
+
 namespace Bit.Core.Services;
 
 public class UserService : UserManager<User>, IUserService, IDisposable
@@ -223,59 +225,39 @@ public class UserService : UserManager<User>, IUserService, IDisposable
 
     public override async Task<IdentityResult> DeleteAsync(User user)
     {
-        // Check if user is the only owner of any organizations.
-        var onlyOwnerCount = await _organizationUserRepository.GetCountByOnlyOwnerAsync(user.Id);
-        if (onlyOwnerCount > 0)
+        var userList = await ValidateDeleteUsersAsync(new List<User> { user });
+        var results = userList.ToList()[0].ErrorMessage;
+
+        if (results == string.Empty || results == null)
         {
-            var deletedOrg = false;
-            var orgs = await _organizationUserRepository.GetManyDetailsByUserAsync(user.Id,
-                OrganizationUserStatusType.Confirmed);
-            if (orgs.Count == 1)
-            {
-                var org = await _organizationRepository.GetByIdAsync(orgs.First().OrganizationId);
-                if (org != null && (!org.Enabled || string.IsNullOrWhiteSpace(org.GatewaySubscriptionId)))
-                {
-                    var orgCount = await _organizationUserRepository.GetCountByOrganizationIdAsync(org.Id);
-                    if (orgCount <= 1)
-                    {
-                        await _organizationRepository.DeleteAsync(org);
-                        deletedOrg = true;
-                    }
-                }
-            }
-
-            if (!deletedOrg)
-            {
-                return IdentityResult.Failed(new IdentityError
-                {
-                    Description = "Cannot delete this user because it is the sole owner of at least one organization. Please delete these organizations or upgrade another user.",
-                });
-            }
+            await _userRepository.DeleteAsync(user);
+            await _referenceEventService.RaiseEventAsync(
+                new ReferenceEvent(ReferenceEventType.DeleteAccount, user, _currentContext));
+            await _pushService.PushLogOutAsync(user.Id);
+            return IdentityResult.Success;
         }
-
-        var onlyOwnerProviderCount = await _providerUserRepository.GetCountByOnlyOwnerAsync(user.Id);
-        if (onlyOwnerProviderCount > 0)
+        else
         {
             return IdentityResult.Failed(new IdentityError
             {
-                Description = "Cannot delete this user because it is the sole owner of at least one provider. Please delete these providers or upgrade another user.",
+                Description = results
             });
         }
+    }
 
-        if (!string.IsNullOrWhiteSpace(user.GatewaySubscriptionId))
+    public async Task<IEnumerable<(Guid UserId, string? ErrorMessage)>> DeleteManyAsync(IEnumerable<User> users)
+    {
+        var results = await ValidateDeleteUsersAsync(users.ToList());
+
+        await _userRepository.DeleteManyAsync(users);
+        foreach (var user in users)
         {
-            try
-            {
-                await CancelPremiumAsync(user);
-            }
-            catch (GatewayException) { }
+            await _referenceEventService.RaiseEventAsync(
+                new ReferenceEvent(ReferenceEventType.DeleteAccount, user, _currentContext));
+            await _pushService.PushLogOutAsync(user.Id);
         }
 
-        await _userRepository.DeleteAsync(user);
-        await _referenceEventService.RaiseEventAsync(
-            new ReferenceEvent(ReferenceEventType.DeleteAccount, user, _currentContext));
-        await _pushService.PushLogOutAsync(user.Id);
-        return IdentityResult.Success;
+        return results;
     }
 
     public async Task<IdentityResult> DeleteAsync(User user, string token)
@@ -286,6 +268,61 @@ public class UserService : UserManager<User>, IUserService, IDisposable
         }
 
         return await DeleteAsync(user);
+    }
+
+    private async Task<IEnumerable<(Guid UserId, string? ErrorMessage)>> ValidateDeleteUsersAsync(List<User> users)
+    {
+        var userResult = new List<(Guid UserId, string? ErrorMessage)>();
+
+        foreach (var user in users)
+        {
+            // Check if user is the only owner of any organizations.
+            var onlyOwnerCount = await _organizationUserRepository.GetCountByOnlyOwnerAsync(user.Id);
+            if (onlyOwnerCount > 0)
+            {
+                var deletedOrg = false;
+                var orgs = await _organizationUserRepository.GetManyDetailsByUserAsync(user.Id,
+                    OrganizationUserStatusType.Confirmed);
+                if (orgs.Count == 1)
+                {
+                    var org = await _organizationRepository.GetByIdAsync(orgs.First().OrganizationId);
+                    if (org != null && (!org.Enabled || string.IsNullOrWhiteSpace(org.GatewaySubscriptionId)))
+                    {
+                        var orgCount = await _organizationUserRepository.GetCountByOrganizationIdAsync(org.Id);
+                        if (orgCount <= 1)
+                        {
+                            await _organizationRepository.DeleteAsync(org);
+                            deletedOrg = true;
+                        }
+                    }
+                }
+
+                if (!deletedOrg)
+                {
+                    userResult.Add((user.Id, "Cannot delete this user because it is the sole owner of at least one organization. Please delete these organizations or upgrade another user."));
+                    continue;
+                }
+            }
+
+            var onlyOwnerProviderCount = await _providerUserRepository.GetCountByOnlyOwnerAsync(user.Id);
+            if (onlyOwnerProviderCount > 0)
+            {
+                userResult.Add((user.Id, "Cannot delete this user because it is the sole owner of at least one provider. Please delete these providers or upgrade another user."));
+                continue;
+            }
+
+            if (!string.IsNullOrWhiteSpace(user.GatewaySubscriptionId))
+            {
+                try
+                {
+                    await CancelPremiumAsync(user);
+                }
+                catch (GatewayException) { }
+            }
+
+            userResult.Add((user.Id, string.Empty));
+        }
+        return userResult;
     }
 
     public async Task SendDeleteConfirmationAsync(string email)
